Add file check for imperative vscode settings

+++ hm/common/modules/Code/xdg.nix
@@ -12,20 +12,24 @@
     before = ["checkLinkTargets"];
     data = ''
       userDir=${config.xdg.configHome}/Code/User
-      rm -rf $userDir/settings.json
+      if [ -if "$userDir"];
+        rm -rf $userDir/settings.json
+      fi
     '';
   };
 
-  #home.activation.afterWriteBoundary = {
-   # after = ["writeBoundary"];
-    #before = [];
-    #data = ''
-    #  userDir=${config.xdg.configHome}/Code/User
-    #  rm -rf $userDir/settings.json
-    #  cat \
-    #    ${(pkgs.formats.json {}).generate "blabla"
-    #    config.programs.vscode.userSettings} \
-    #    > $userDir/settings.json
-    #'';
-  #};
+  home.activation.afterWriteBoundary = {
+    after = ["writeBoundary"];
+    before = [];
+    data = ''
+      userDir=${config.xdg.configHome}/Code/User
+      if [ -if "$userDir"];
+        rm -rf $userDir/settings.json
+      fi
+      cat \
+        ${(pkgs.formats.json {}).generate "blabla"
+        config.programs.vscode.userSettings} \
+        > $userDir/settings.json
+    '';
+  };
 }
